+++ seleniumpm/config.py
@@ -1,0 +1,8 @@
+page_timeout_in_sec = 30
+element_timeout_in_sec = 10
+
+def reset():
+    global page_timeout_in_sec
+    global element_timeout_in_sec
+    page_timeout_in_sec = 30
+    element_timeout_in_sec = 10