+++ HISTORY.rst
@@ -9,7 +9,6 @@ FUTURE (UN-RELEASED)
 - Added take_screenshot() feature to a Webpage. Also added a take_screenshot_on_error annotation for annotating tests.
 - Added pretty-print way of visualizing all attributes and methods for a Webpage. This is useful for documentation as well as for debugging
 - Added highlight() feature to all Element types
-- Added a global page_element_timeout and page_load_timeout for all of SeleniumPM. Defaults are still 10s for page_element_timeout and 30s for page_load_timeout
 - Added ability to calculate durations between common operations (e.g. page or widget load times)
 
 **Fixed:**
@@ -22,6 +21,8 @@ FUTURE (UN-RELEASED)
 **New Features:**
 
 - Added a get_attribute_contains() and a get_attribute_is() on Element type. The intention is to implement an all-in-one solution for both retrieving an attribute and validating whether something contains or is something. This operation is useful for dynamic elements that use css classes for changing the state of a page (e.g. <div class='svgOverviewView-status-icon fa fa-exclamation-triangle svg-status-warning'>); One could define a generic reference to this element, and then use get_attribute_contains() to check if it now contains 'fa-exclamation-triangle' css class.
+- Implementation of seleniumpm.config module. This module can be imported via *import seleniumpm.config as seleniumconfig*; and variables can be changed like so: *seleniumconfig.page_timeout_in_sec = 60*
+- Added a global element_timeout_in_ms and page_timeout_in_ms to seleniumpm.config module. Defaults are still 10s for element timeout and 30s for page timeout.
 - Added a refresh() operation on a Webpage (a.k.a. a page refresh). This of course is simply a driver.refresh() operation on an opened page.
 - Added a get_current_url() operation on a Webpage
 - Added get_number(), get_numbers(), get_int(), and get_float() to an Element. This is so that you can quickly and easily get numbers from element(s) text that represent numbers